requirements.txt
opcua
cryptography
python-dotenv
paho-mqtt
